Refactor mooc_backend app folder
