+++ cucumber.js
@@ -9,8 +9,8 @@ const backoffice_backend = [
 ].join(' ');
 const mooc_backend = [
   ...common,
-  'tests/apps/mooc_backend/features/**/*.feature',
-  '--require tests/apps/mooc_backend/features/step_definitions/*.steps.ts'
+  'tests/apps/mooc/backend/features/**/*.feature',
+  '--require tests/apps/mooc/backend/features/step_definitions/*.steps.ts'
 ].join(' ');
 
 module.exports = {
