+++ src/Contexts/Backoffice/Courses/infrastructure/persistence/BackofficeMongoConfigFactory.ts
@@ -1,4 +1,4 @@
-import config from '../../../../../apps/mooc_backend/config/config';
+import config from '../../../../../apps/mooc/backend/config/config';
 import MongoConfig from '../../../../Shared/infrastructure/persistence/mongo/MongoConfig';
 
 export class BackofficeMongoConfigFactory {
