+++ src/apps/mooc/backend/config/config.ts
