+++ src/apps/mooc/backend/config/dependency-injection/Courses/application.yaml
@@ -1,15 +1,15 @@
 services:
 
   Mooc.courses.CourseRepository:
-    class: ../../../../../Contexts/Mooc/Courses/infrastructure/persistence/MongoCourseRepository
+    class: ../../../../../../Contexts/Mooc/Courses/infrastructure/persistence/MongoCourseRepository
     arguments: ['@Shared.ConnectionManager']
 
   Mooc.courses.CourseCreator:
-    class: ../../../../../Contexts/Mooc/Courses/application/CourseCreator
+    class: ../../../../../../Contexts/Mooc/Courses/application/CourseCreator
     arguments: ['@Mooc.courses.CourseRepository', '@Shared.EventBus']
 
   Mooc.courses.CreateCourseCommandHandler:
-    class: ../../../../../Contexts/Mooc/Courses/application/CreateCourseCommandHandler
+    class: ../../../../../../Contexts/Mooc/Courses/application/CreateCourseCommandHandler
     arguments: ['@Mooc.courses.CourseCreator']
     tags:
       - { name: 'commandHandler' }
