+++ src/apps/mooc/backend/config/dependency-injection/application.yaml
