+++ src/apps/mooc/backend/config/dependency-injection/application_production.yaml
