+++ src/apps/mooc/backend/config/dependency-injection/application_test.yaml
@@ -1,0 +1,11 @@
+imports:
+  - { resource: ./application.yaml }
+
+services:
+  Mooc.EnvironmentArranger:
+    class: ../../../../../../tests/Contexts/Shared/infrastructure/mongo/MongoEnvironmentArranger
+    arguments: ['@Shared.ConnectionManager']
+
+  Shared.EventBus:
+    class: ../../../../../Contexts/Shared/infrastructure/EventBus/InMemorySyncEventBus
+    arguments: []
