+++ src/apps/mooc/backend/config/dependency-injection/apps/application.yaml
