+++ src/apps/mooc/backend/config/dependency-injection/index.ts
