+++ src/apps/mooc/backend/controllers/CoursePutController.ts
@@ -1,9 +1,9 @@
 import { Request, Response } from 'express';
 import httpStatus from 'http-status';
 import { Controller } from './Controller';
-import { CourseAlreadyExists } from '../../../Contexts/Mooc/Courses/domain/CourseAlreadyExists';
-import { CommandBus } from '../../../Contexts/Shared/domain/CommandBus';
-import { CreateCourseCommand } from '../../../Contexts/Mooc/Courses/application/CreateCourseCommand';
+import { CourseAlreadyExists } from '../../../../Contexts/Mooc/Courses/domain/CourseAlreadyExists';
+import { CommandBus } from '../../../../Contexts/Shared/domain/CommandBus';
+import { CreateCourseCommand } from '../../../../Contexts/Mooc/Courses/application/CreateCourseCommand';
 
 export class CoursePutController implements Controller {
   constructor(private commandBus: CommandBus) {}
