+++ src/apps/mooc/backend/routes/courses.route.ts
