+++ src/apps/mooc/backend/subscribers.ts
@@ -1,8 +1,8 @@
 import container from './config/dependency-injection';
-import { InMemoryAsyncEventBus } from '../../Contexts/Shared/infrastructure/EventBus/InMemoryAsyncEventBus';
+import { InMemoryAsyncEventBus } from '../../../Contexts/Shared/infrastructure/EventBus/InMemoryAsyncEventBus';
 import { Definition } from 'node-dependency-injection';
-import { DomainEventSubscriber } from '../../Contexts/Shared/domain/DomainEventSubscriber';
-import { DomainEvent } from '../../Contexts/Shared/domain/DomainEvent';
+import { DomainEventSubscriber } from '../../../Contexts/Shared/domain/DomainEventSubscriber';
+import { DomainEvent } from '../../../Contexts/Shared/domain/DomainEvent';
 
 export function registerSubscribers() {
   const eventBus = container.get('Shared.EventBus') as InMemoryAsyncEventBus;
