--- src/apps/mooc_backend/config/dependency-injection/Shared/application.yaml
@@ -1,43 +1,0 @@
-services:
-  Shared.MongoConfig:
-    factory:
-      class: ../../../../../Contexts/Mooc/Shared/infrastructure/persistence/mongo/MongoConfigFactory
-      method: 'createConfig'
-
-  Shared.ConnectionManager:
-    factory:
-      class: ../../../../../Contexts/Shared/infrastructure/persistence/mongo/MongoClientFactory
-      method: 'createClient'
-    arguments: ['mooc', '@Shared.MongoConfig']
-
-  Shared.Logger:
-    class: ../../../../../Contexts/Shared/infrastructure/WinstonLogger
-    arguments: []
-
-  Shared.EventBus:
-    class: ../../../../../Contexts/Shared/infrastructure/EventBus/InMemoryAsyncEventBus
-    arguments: []
-
-  Shared.CommandHandlersInformation:
-    class: ../../../../../Contexts/Shared/infrastructure/CommandBus/CommandHandlersInformation
-    arguments: ['!tagged commandHandler']
-
-  Shared.CommandBus:
-    class: ../../../../../Contexts/Shared/infrastructure/CommandBus/InMemoryCommandBus
-    arguments: ['@Shared.CommandHandlersInformation']
-
-  Shared.EventBus.DomainEventMapping:
-    class: ../../../../../Contexts/Shared/infrastructure/EventBus/DomainEventMapping
-    arguments: ['!tagged domainEventSubscriber']
-
-  Shared.EventBus.DomainEventJsonDeserializer:
-    class: ../../../../../Contexts/Shared/infrastructure/EventBus/DomainEventJsonDeserializer
-    arguments: ['@Shared.EventBus.DomainEventMapping']
-
-  Shared.QueryHandlersInformation:
-    class: ../../../../../Contexts/Shared/infrastructure/QueryBus/QueryHandlersInformation
-    arguments: ['!tagged queryHandler']
-
-  Shared.QueryBus:
-    class: ../../../../../Contexts/Shared/infrastructure/QueryBus/InMemoryQueryBus
-    arguments: ['@Shared.QueryHandlersInformation']
