+++ tests/apps/mooc/backend/features/courses_counter/get-courses-counter.feature
