+++ tests/utils/cypress/plugins/index.ts
@@ -1,7 +1,7 @@
 import backendBackofficecontainer from '../../../../src/apps/backoffice/backend/config/dependency-injection';
 import { registerSubscribers } from '../../../../src/apps/backoffice/backend/subscribers';
 import { seed } from '../../../../src/apps/backoffice/frontend/seed';
-import moocContainer from '../../../../src/apps/mooc_backend/config/dependency-injection';
+import moocContainer from '../../../../src/apps/mooc/backend/config/dependency-injection';
 import { EventBus } from '../../../../src/Contexts/Shared/domain/EventBus';
 import { DomainEventJsonDeserializer } from '../../../../src/Contexts/Shared/infrastructure/EventBus/DomainEventJsonDeserializer';
 import { EnvironmentArranger } from '../../../Contexts/Shared/infrastructure/arranger/EnvironmentArranger';
